requirements.txt
appdirs==1.4.3
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
cryptography==2.8
distlib==0.3.0
Django==2.2.5
django-mongoengine==0.4.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
elastic-apm==5.3.3
entrypoints==0.3
Faker==4.0.1
filelock==3.0.12
flake8==3.7.9
fuzzywuzzy==0.18.0
geographiclib==1.50
geopy==1.22.0
identify==1.4.13
idna==2.8
importlib-metadata==1.5.0
mccabe==0.6.1
mongoengine==0.19.1
more-itertools==8.2.0
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
nodeenv==1.3.5
packaging==20.3
pi==0.1.2
Pillow==7.0.0
pluggy==0.13.1
pre-commit==2.2.0
pre-commit-hooks==2.5.0
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
pyflakes==2.1.1
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.2
PyYAML==5.3.1
requests==2.22.0
routing==0.2.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Shapely==1.7.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.0
urllib3==1.25.8
virtualenv==20.0.15
wcwidth==0.1.8
zipp==3.1.0
